Permit admin to change and remove ads or comments.

+++ src/main/java/ru/digitalmagicians/ebito/service/impl/AdsServiceImpl.java
@@ -14,8 +14,10 @@ import ru.digitalmagicians.ebito.entity.Ads;
 import ru.digitalmagicians.ebito.entity.Image;
 import ru.digitalmagicians.ebito.exception.AdsValidationException;
 import ru.digitalmagicians.ebito.exception.CommentNotFoundException;
+import ru.digitalmagicians.ebito.exception.PermissionDeniedException;
 import ru.digitalmagicians.ebito.mapper.AdsMapper;
 import ru.digitalmagicians.ebito.repository.AdsRepository;
+import ru.digitalmagicians.ebito.security.AccessChecker;
 import ru.digitalmagicians.ebito.service.AdsService;
 import ru.digitalmagicians.ebito.service.ImageService;
 import ru.digitalmagicians.ebito.service.UserService;
@@ -34,6 +36,7 @@ public class AdsServiceImpl implements AdsService {
     private final UserService userService;
     private final AdsMapper adsMapper;
     private final ImageService imageService;
+    private final AccessChecker accessChecker;
 
     @Override
     public AdsDto createAds(MultipartFile image, CreateAdsDto properties, Authentication authentication) {
@@ -60,19 +63,22 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public AdsDto updateAds(Integer id, CreateAdsDto createAds) {
-        if (validation(createAds)) {
-            log.error("empty fields CreateAdsDto updateAds");
-            throw new AdsValidationException("empty fields updateAds");
-        }
+    public AdsDto updateAds(Integer id, CreateAdsDto createAds){
         Ads ads = getAdsById(id);
-        ads.setTitle(createAds.getTitle());
-        ads.setDescription(createAds.getDescription());
-        ads.setPrice(createAds.getPrice());
-        Ads savedAds = adsRepository.save(ads);
-        log.info("Successful updating ads by id: {}", id);
-
-        return adsMapper.toDto(savedAds);
+        if (accessChecker.checkAccess(ads)){
+            if (validation(createAds)) {
+                log.error("empty fields CreateAdsDto updateAds");
+                throw new AdsValidationException("empty fields updateAds");
+            }
+            ads.setTitle(createAds.getTitle());
+            ads.setDescription(createAds.getDescription());
+            ads.setPrice(createAds.getPrice());
+            Ads savedAds = adsRepository.save(ads);
+            log.info("Successful updating ads by id: {}", id);
+            return adsMapper.toDto(savedAds);
+        } else {
+            throw new PermissionDeniedException();
+        }
     }
 
     @Override
@@ -115,8 +121,12 @@ public class AdsServiceImpl implements AdsService {
     @Override
     public void delete(Integer id) {
         Ads ads = getAdsById(id);
-        adsRepository.delete(ads);
-        log.info("Successful deleting ads by id: {}", id);
+        if (accessChecker.checkAccess(ads)) {
+            adsRepository.delete(ads);
+            log.info("Successful deleting ads by id: {}", id);
+        } else {
+            throw new PermissionDeniedException();
+        }
     }
 
     @Override
